refactor: move and rename S components
