+++ apps/ui/src/components/Modal/EditContact.vue
@@ -79,7 +79,7 @@ watch(
       <h3 v-text="initialState?.address ? 'Edit contact' : 'Add contact'" />
     </template>
     <div class="s-box p-4">
-      <SIObject :model-value="form" :error="formErrors" :definition="definition" />
+      <UiForm :model-value="form" :error="formErrors" :definition="definition" />
     </div>
     <template #footer>
       <UiButton class="w-full" :disabled="Object.keys(formErrors).length > 0" @click="handleSubmit">
