+++ apps/ui/src/components/Ui/InputNumber.vue
@@ -1,11 +1,5 @@
-<script lang="ts">
-export default {
-  inheritAttrs: false
-};
-</script>
-
 <script setup lang="ts">
-const model = defineModel<string>();
+const model = defineModel<string | number>();
 
 const props = defineProps<{
   error?: string;
@@ -34,12 +28,13 @@ watch(model, () => {
 </script>
 
 <template>
-  <SBase :definition="definition" :error="error" :dirty="dirty">
-    <textarea
+  <UiWrapperInput :definition="definition" :error="error" :dirty="dirty">
+    <input
       v-model="inputValue"
+      type="number"
       class="s-input"
       v-bind="$attrs"
       :placeholder="definition.examples && definition.examples[0]"
     />
-  </SBase>
+  </UiWrapperInput>
 </template>
