+++ apps/ui/src/components/Ui/InputStamp.vue
