+++ apps/ui/src/components/Ui/Select.vue
@@ -1,65 +1,49 @@
-<script setup lang="ts" generic="T extends string | number, U extends readonly Item<T>[]">
-import type { Component } from 'vue';
-
-export type Item<T extends string | number> = {
-  key: T;
-  label: string;
-  indicator?: string;
-  component?: Component;
-  componentProps?: Record<string, unknown>;
+<script lang="ts">
+export default {
+  inheritAttrs: false
 };
+</script>
 
-const model = defineModel<U[number]['key']>({ required: true });
+<script setup lang="ts">
+const model = defineModel<string>();
 
 const props = defineProps<{
-  title: string;
-  items: U;
-  gap?: string;
-  placement?: 'left' | 'right';
+  error?: string;
+  definition: any;
 }>();
 
-const currentItem = computed(() => props.items.find(item => item.key === model.value));
-const items = computed(() => props.items);
+const dirty = ref(false);
+
+const inputValue = computed({
+  get() {
+    if (!model.value && !dirty.value && props.definition.default) {
+      return props.definition.default;
+    }
+
+    return model.value;
+  },
+  set(newValue: string) {
+    dirty.value = true;
+    model.value = newValue;
+  }
+});
+
+watch(model, () => {
+  dirty.value = true;
+});
 </script>
 
 <template>
-  <UiDropdown :gap="gap" :placement="placement">
-    <template #button>
-      <slot name="button">
-        <button
-          class="flex items-center gap-2 relative rounded-full leading-[100%] border button px-[16px] min-w-[76px] h-[42px] top-1 outline-0 text-skin-link bg-skin-bg"
-        >
-          <div class="absolute top-[-10px] bg-skin-bg px-1 left-[12px] text-[16px] text-skin-text">
-            {{ title }}
-          </div>
-          <template v-if="currentItem">
-            <div
-              v-if="currentItem.indicator"
-              class="w-[8px] h-[8px] rounded-full"
-              :class="currentItem.indicator"
-            />
-            <component
-              :is="currentItem.component"
-              v-else-if="currentItem.component"
-              v-bind="currentItem.componentProps"
-            />
-            {{ currentItem.label }}
-          </template>
-        </button>
-      </slot>
-    </template>
-    <template #items>
-      <UiDropdownItem v-for="item in items" :key="item.key" v-slot="{ active, disabled }">
-        <button
-          class="flex items-center gap-2"
-          :class="{ 'opacity-80': active, 'opacity-40': disabled }"
-          @click="model = item.key"
-        >
-          <div v-if="item.indicator" class="w-[8px] h-[8px] rounded-full" :class="item.indicator" />
-          <component :is="item.component" v-else-if="item.component" v-bind="item.componentProps" />
-          {{ item.label }}
-        </button>
-      </UiDropdownItem>
-    </template>
-  </UiDropdown>
+  <UiWrapperInput :definition="definition" :error="error" :dirty="dirty">
+    <select v-model="inputValue" class="s-input">
+      <option disabled value="">Please select one</option>
+      <option
+        v-for="option in definition.options || definition.enum"
+        :key="option.id === undefined ? option : option.id"
+        :value="option.id === undefined ? option : option.id"
+      >
+        {{ option.name ?? option }}
+      </option>
+    </select>
+  </UiWrapperInput>
 </template>
